routing, payments, deleted files that were causing errors, left kellys profile and userInformation, fixed file structure.
