+++ LinkedCodes/app/(tabs)/_layout.jsx
@@ -1,4 +1,4 @@
-/*import React from "react";
+import React from "react";
 import { Tabs } from "expo-router";
 import Ionicons from "@expo/vector-icons/Ionicons";
 import FontAwesome6 from "@expo/vector-icons/FontAwesome6";
@@ -6,16 +6,16 @@ import FontAwesome6 from "@expo/vector-icons/FontAwesome6";
 const Layout = () => {
   return (
     <Tabs
-    screenOptions={{
-      headerShown: false,
-      tabBarActiveTintColor: '#202A44', 
-      tabBarInactiveTintColor: 'gray',
-      tabBarStyle: {
-        display: 'flex', 
-        backgroundColor: 'white',
-      },
-    }}
-    initialRouteName='Home'
+      screenOptions={{
+        headerShown: false,
+        tabBarActiveTintColor: '#202A44',
+        tabBarInactiveTintColor: 'gray',
+        tabBarStyle: {
+          display: 'flex',
+          backgroundColor: 'white',
+        },
+      }}
+      initialRouteName='Home'
     >
       <Tabs.Screen
         name="Home"
@@ -28,7 +28,7 @@ const Layout = () => {
       <Tabs.Screen
         name="Maintainance"
         options={{
-          tabBarIcon: ({ color  }) => (
+          tabBarIcon: ({ color }) => (
             <Ionicons name="cog-outline" color={color} size={34} />
           ),
         }}
@@ -49,66 +49,13 @@ const Layout = () => {
       <Tabs.Screen
         name="Settings"
         options={{
-          tabBarIcon: ({ color}) => (
+          tabBarIcon: ({ color }) => (
             <Ionicons name="settings-outline" color={color} size={30} />
           ),
         }}
       />
     </Tabs>
   );
-};
-
-export default Layout;
-*/
-import { Tabs } from "expo-router";
-import Ionicons from "@expo/vector-icons/Ionicons";
-import FontAwesome6 from "@expo/vector-icons/FontAwesome6";
-
-const Layout = () => {
-    return (
-        <Tabs
-            initialRouteName="home" // Ensure this matches the name used in the Home component
-            screenOptions={{
-                headerShown: false,
-                tabBarActiveTintColor: '#202A44',
-                tabBarInactiveTintColor: 'gray',
-                tabBarStyle: {
-                    display: 'flex',
-                    backgroundColor: 'white',
-                },
-            }}
-        >
-            <Tabs.Screen
-                name="home" // Ensure this matches the Home component name
-                options={{
-                    tabBarIcon: ({ color, size }) => (
-                        <Ionicons name="home-outline" color={color} size={size} />
-                    ),
-                }}
-            />
-            <Tabs.Screen
-                name="reporting"
-                options={{
-                    tabBarIcon: ({ color }) => (
-                        <Ionicons name="clipboard-outline" color={color} size={34} />
-                    ),
-                }}
-            />
-            <Tabs.Screen
-                name="traffic"
-                options={{
-                    tabBarIcon: ({ color, size }) => (
-                        <FontAwesome6
-                            name="map"
-                            color={color}
-                            size={size}
-                            stroke={0.9}
-                        />
-                    ),
-                }}
-            />
-        </Tabs>
-    );
 };
 
 export default Layout;
