+++ LinkedCodes/app/(userTabs)/home/_layout.jsx
@@ -1,5 +1,3 @@
-import { StyleSheet, Text, View } from 'react-native'
-import React from 'react'
 import { StatusBar } from 'expo-status-bar'
 import { Stack } from 'expo-router'
 
@@ -7,7 +5,7 @@ const Layout = () => {
     return (
         <>
             <StatusBar style="light" />
-            <Stack screenOptions={{ headerShown: false}} />
+            <Stack screenOptions={{ headerShown: false }} />
         </>
     )
 }
