+++ LinkedCodes/app/(userTabs)/reporting/_layout.jsx
@@ -1,0 +1,3 @@
+import { Stack } from "expo-router";
+
+export default Stack;
