+++ LinkedCodes/package.json
@@ -25,6 +25,7 @@
     "@types/react": "~18.2.79",
     "axios": "^1.7.7",
     "expo": "^51.0.32",
+    "expo-camera": "^15.0.16",
     "expo-constants": "~16.0.2",
     "expo-contacts": "~13.0.5",
     "expo-device": "~6.0.2",
@@ -46,6 +47,8 @@
     "react-native": "0.74.5",
     "react-native-biometrics": "^3.0.1",
     "react-native-calendars": "^1.1306.0",
+    "react-native-drawer-layout": "^3.3.2",
+    "react-native-geocoding": "^0.5.0",
     "react-native-gesture-handler": "~2.16.1",
     "react-native-gifted-chat": "^2.6.4",
     "react-native-image-picker": "^7.1.2",
@@ -60,12 +63,16 @@
     "react-native-super-grid": "^6.0.1",
     "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.10",
-    "react-step-builder": "^3.0.6"
+    "react-step-builder": "^3.0.6",
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.24.0",
     "tailwindcss": "^3.3.2",
     "typescript": "~5.3.3"
   },
-  "private": true
+  "private": true,
+  "trustedDependencies": [
+    "protobufjs"
+  ]
 }
