+++ LinkedCodes/src/hamburger/emergency.jsx
