+++ LinkedCodes/src/profile/firebase.jsx
@@ -1,0 +1,2 @@
+import { auth, db, storage } from "../../firebase"
+export { auth, db, storage };
