--- package.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": {
-    "@types/react": "~18.2.79",
-    "expo": "^51.0.37",
-    "metro": "^0.80.12",
-    "react-native": "^0.74.5",
-    "typescript": "~5.3.3"
-  },
-  "devDependencies": {
-    "@expo/cli": "^0.18.30"
-  }
-}
